Agreeing with Robert Watson (bugreport #281) that it is unlikely that random() was the cause of memory corruption, it is more likely that something passed an uninitialized pointer to the RNG state table as an argument to rpc2_NextRandom. Closing the bugreport and removing half of the random number generation code.

+++ lib-src/rpc2/include/rpc2/rpc2.h
@@ -750,7 +750,7 @@ extern void rpc2_Encrypt (char *FromBuffer, char *ToBuffer,
 
 void rpc2_Decrypt (char *FromBuffer, char *ToBuffer, long  HowManyBytes,
     RPC2_EncryptionKey WhichKey, int EncryptionType);
-extern long rpc2_NextRandom (char *StatePtr);
+extern unsigned int rpc2_NextRandom (char *StatePtr);
 
 /* hack until we can do something more sophisticated. */
 extern long rpc2_Bandwidth;
